+++ CHANGELOG.md
@@ -183,6 +183,7 @@
 
 [0.1.1]: https://github.com/TNick/roborean/compare/846ef574a1c267a96143a80cb33ffe8e935d737c...v0.1.1
 [0.1.10]: https://github.com/TNick/roborean/compare/v0.1.9...v0.1.10
+[0.1.11]: https://github.com/TNick/roborean/compare/v0.1.10...v0.1.11
 [0.1.2]: https://github.com/TNick/roborean/compare/v0.1.1...v0.1.2
 [0.1.3]: https://github.com/TNick/roborean/compare/v0.1.2...v0.1.3
 [0.1.4]: https://github.com/TNick/roborean/compare/v0.1.3...v0.1.4
@@ -191,4 +192,4 @@
 [0.1.7]: https://github.com/TNick/roborean/compare/v0.1.6...v0.1.7
 [0.1.8]: https://github.com/TNick/roborean/compare/v0.1.7...v0.1.8
 [0.1.9]: https://github.com/TNick/roborean/compare/v0.1.8...v0.1.9
-[unreleased]: https://github.com/TNick/roborean/compare/v0.1.10...HEAD
+[unreleased]: https://github.com/TNick/roborean/compare/v0.1.11...HEAD